requirements.txt
accelerate
datasets
hydra-core
omegaconf
openai
rouge_score
torch
transformers
wandb
bitsandbytes
peft
torch
fairscale
scipy
fire
sentencepiece
trl
pyext
deepspeed
matplotlib